Create C2 not working
I tried using the C2 to perform a controlled explotion on a remote bomb but it did not work. No matter what I did, it would not put the C2 on the bomb. Several people tried to put C2 on it, it didn't work which in the end blew up by someone calling it.